🚗 Step3: 자동차 경주 (#1230)
* refactor(Expression): 기본 구분자 상수화

* refactor(Operator): 연산자 못찾을 시 커스텀 예외 발생

* refactor: 계산시 BigDecimal 타입 사용하도록 변경

* feat(Expression): 식의 숫자가 잘못된 경우 예외가 발생한다

* refactor(OperatorTest): operator 예약어 사용하는 변수명 변경

* feat: 계산기 실행 main 함수 추가

* docs: 자동차 경주 요구사항 작성

* feat(Car): 자동차는 위치를 갖는다

* feat(Car): 자동차의 기본 위치는 0이다

* feat(Car): 자동차는 조건에 따라 전진 또는 정지한다

* refactor: 전략, 숫자 생성기 분리

* feat(Cars): 자동차는 여러 대가 존재한다

* feat(Cars): 자동차는 각각 전진 또는 정지한다

* refactor: 패키지 분리

* feat(RacingGame): 레이싱 게임 구현

* feat: 레이싱 게임 클라이언트 구현

* refactor(Expression): 숫자 변환 한 번만 시도하도록 변경

* refactor(Main): main 파라미터 제거

* refactor(Operator): 계산시 함수 호출 형태로 변경

* refactor(Car, Cars): 자동차 생성시 이동 전략 주입

* refactor(Operator): java stream 대신 kotlin sequence 사용하도록 변경

* refactor(RoundHistory): trailing comma 추가

* refactor(RandomNumberGenerator): IntRange 저장 후 재사용

+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,7 +1,9 @@
 package calculator
 
+import java.math.BigDecimal
+
 class Calculator {
-    fun calculate(expression: String): Int {
+    fun calculate(expression: String): BigDecimal {
         return Expression.from(expression).getResult()
     }
 }
